generate signatures for channel metadata with releasekey
For RSA keys, we sign using RSASSA-PSS as recommended for new
applications in PKCS #1 v2.2 (RFC 8017) [1].  The default saltLength for
the RSASSA-PSS-params is the octet length of the hash value, so we use
the special "digest" value for `rsa_pss_saltlen` in order to be explicit
[2].  SHA256withRSA/PSS is supported on Android [3].

ECDSA keys are also supported, as they can be generated from the
`development/tools/make_key` script. SHA256withECDSA is supported on
Android.

Test: Ran through the process of `m target-files-package` and
`./release.sh sargo`, and signature verification, update download, and
update installation checked with the Updater app on a self-hosted
server for RSA-2048, RSA-4096, and ECDSA keys.

[1] https://tools.ietf.org/html/rfc8017#section-8
[2] https://www.openssl.org/docs/man1.1.1/man1/openssl-pkeyutl.html#RSA-ALGORITHM
[3] https://developer.android.com/reference/java/security/Signature