switch to new Pixel kernel repository system
The most recent Pixel kernels tags pushed to AOSP were for the initial
release of Android 16. We're now obtaining Pixel kernel sources in the
form of tarballs without Git history which we put into a Git repository
of our own to track the history of upstream changes in a structured way
with our changes rebased on top of it. We replace the aosp/ tree within
the source tree with our common kernel repository to provide the Generic
Kernel Image. We're using a Git submodule for this rather than repo for
now so we don't currently have a kernel manifest repository for Pixels.

This whole situation is far from ideal and we'd greatly prefer if Pixel
kernels were once again pushed as branches/tags with commit history to a
bunch of repositories in AOSP.